refactor: clean up setting module and introduce admin OpenAPI endpoints
