+++ apps/nestjs-backend/src/features/setting/open-api/admin-open-api.module.ts
@@ -1,0 +1,21 @@
+import { Module } from '@nestjs/common';
+import { MulterModule } from '@nestjs/platform-express';
+import multer from 'multer';
+import { AttachmentsCropModule } from '../../attachments/attachments-crop.module';
+import { StorageModule } from '../../attachments/plugins/storage.module';
+import { AdminOpenApiController } from './admin-open-api.controller.ts';
+import { AdminOpenApiService } from './admin-open-api.service.ts';
+
+@Module({
+  imports: [
+    AttachmentsCropModule,
+    MulterModule.register({
+      storage: multer.diskStorage({}),
+    }),
+    StorageModule,
+  ],
+  controllers: [AdminOpenApiController],
+  exports: [AdminOpenApiService],
+  providers: [AdminOpenApiService],
+})
+export class AdminOpenApiModule {}
